+++ CHANGELOG.rst
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 .. contents:: Table of Contents
    :depth: 2
 
+Unreleased
+==========
+
+Added
+-----
+* Add autofocus to the username input on the login page
+
+
 v0.7.2 - 2016-08-31
 ===================